+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## 0.207.1
+
+- Fix email parser.
+
 ## 0.207.0
 
 - Change heading and dlist syntax to disable annotation, reference, index, and link syntax.